Resolved conflict in:
	src/pmlogger/events.c

+++ src/libpcp/src/derive_fetch.c
@@ -10,10 +10,6 @@
  * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
  * or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public
  * License for more details.
- * 
- * You should have received a copy of the GNU Lesser General Public License
- * along with this library; if not, write to the Free Software Foundation,
- * Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307 USA.
  *
  * Debug Flags
  *	DERIVE - high-level diagnostics
@@ -22,6 +18,7 @@
  *	DERIVE & APPL2 - fetch handling
  */
 
+#include <inttypes.h>
 #include "derive.h"
 
 static void
@@ -1068,9 +1065,9 @@ __dmpostfetch(__pmContext *ctxp, pmResult **result)
 	    else if (cp->mlist[m].expr->desc.type == PM_TYPE_U32)
 		fprintf(stderr, " u=%u", cp->mlist[m].expr->info->ivlist[k].value.ul);
 	    else if (cp->mlist[m].expr->desc.type == PM_TYPE_64)
-		fprintf(stderr, " ll=%lld", (long long)cp->mlist[m].expr->info->ivlist[k].value.ll);
+		fprintf(stderr, " ll=%"PRIi64, cp->mlist[m].expr->info->ivlist[k].value.ll);
 	    else if (cp->mlist[m].expr->desc.type == PM_TYPE_U64)
-		fprintf(stderr, " ul=%llu", (unsigned long long)cp->mlist[m].expr->info->ivlist[k].value.ull);
+		fprintf(stderr, " ul=%"PRIu64, cp->mlist[m].expr->info->ivlist[k].value.ull);
 	    else if (cp->mlist[m].expr->desc.type == PM_TYPE_FLOAT)
 		fprintf(stderr, " f=%f", (double)cp->mlist[m].expr->info->ivlist[k].value.f);
 	    else if (cp->mlist[m].expr->desc.type == PM_TYPE_DOUBLE)
