+++ src/libpcp/src/units.c
@@ -10,15 +10,12 @@
  * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
  * or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public
  * License for more details.
- * 
- * You should have received a copy of the GNU Lesser General Public License
- * along with this library; if not, write to the Free Software Foundation,
- * Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307 USA.
  */
 
 #include "pmapi.h"
 #include "impl.h"
 #include <math.h>
+#include <inttypes.h>
 
 #ifndef ABS
 #define    ABS(a)          ((a) < 0 ? -(a) : (a))
@@ -50,10 +47,10 @@ pmAtomStr(const pmAtomValue *avp, int type)
 	    snprintf(buf, sizeof(buf), "%u", av.ul);
 	    break;
 	case PM_TYPE_64:
-	    snprintf(buf, sizeof(buf), "%lld", (long long)av.ll);
+	    snprintf(buf, sizeof(buf), "%"PRIi64, av.ll);
 	    break;
 	case PM_TYPE_U64:
-	    snprintf(buf, sizeof(buf), "%llu", (unsigned long long)av.ull);
+	    snprintf(buf, sizeof(buf), "%"PRIu64, av.ull);
 	    break;
 	case PM_TYPE_FLOAT:
 	    snprintf(buf, sizeof(buf), "%e", (double)av.f);
